- Moved logic for parsing properties of an object to DebugUtils.cs - Populated and documented InvalidEventSubscriptionException.cs - Added Try-Catch to ItemManager.cs registration method.
